Settings file
The setting files contains the necessary client ID and passcode necessary for the QR tool to operate. This information is specific to your organization or account. For EPA users, please contact Timothy Boe and Jonathan Pettit to obtain the settings file that contains this information.

The url, client ID, and passcode have been removed from the public version of this file.